Add basic org.gnome.Mutter.Idle DBus protocol support
This allows e.g. g-s-d power track our idle time for e.g. dimming the
display. We will be able to use this for screen locking later on too.

We only implement the core timer since we're not interested in timeouts
for individual monitors atm. The used Wayland protocoll doesn't support
this either.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>