ScreensaverManager: Hook up to LockscreenManager
Implement org.gnome.ScreenSaver by invoking the appropriate methods
on our Lockscreenmanager and listen to it's releveant signals.

We keep this in a separte object to keep the lock screen code
separte from the DBus interaction.

We inject the lock screen manager instead of fetching the singleton
to make the code more self contained.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>