AppGrid: Live update search when using an input method.
When using an input method for text input (as some osks do, and squeekboard may
in the future), the entry doesn't receive individual key strokes. It only
receives the text when a string is fully composed (user hits enter, space,
period, exclam, etc, whatever the input method uses to terminate composing).
This works fine for most text input, but not for live updating on key strokes.

It is for cases like these, that GtkEntry::preedit-changed exists. It gives
access to the buffer the input method uses, as it gets updated.