Drop external margin from wwaninfo
f0a582d3c9d introduced a fixed margin to the right in the wwaninfo
widget that broke alignment on the lock screen. Drop that and use an
always internal margin between the two icons (the access_tex text (which
will become an icon once available) will never be shown without the
signal strenght icon).

Signed-off-by: Guido Günther <guido.gunther@puri.sm>