app-list-model: phantom launcher
At last a fix for the missing item problem

tl;dr: we skipped the first entry

The reason it was weirdly deterministic is that we _always_ skipped
the first item and as long as the entries didn't change the item would
_always_ be the same

Why was the item seemingly random? Because the list doesn't have any
sorting at the point we lost an item

So why didn't this effect more people? A system has many, many entries -
a lot of which aren't actually app launchers. For example my laptop
currently has 301 entries of which 172 should actually be shown. By pure
fluke we got lucky most of the time and skipped an unintersting one.

The fix is painfully simple: iterate the GList properly