background-manager: Notify background about configuration changes
We notify the background when the configuration changes. This avoids
having to leak any monitor details into PhoshMonitor itself.

Ideally we'd later on also pass the panel exclusion area from
PhoshMonitorManager so the background can figure out what to draw
from it's surface properties and other properties we feed it.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>