po: Fix German translation of Phone Shell
Application name "Phone Shell" was translated as "Telefonhülle" which is
definitely something different (think Otterbox). To stay in line with
GNOME Shell which remains untranslated too, I propose to keep "Phone
Shell" as a terminus technicus. "Telefon Shell" is a weird mix of
English and German and does not sound nice either. Fixes #347.