Add a custom log handler
This is to a large extend what glib does but we don't read the levels
from the environment but allow to set them when setting the handler.

Resetting the handler instead of just adjusting the domains has the
advantage that we can pass in new user_data. Otherwise we'd have to keep
the current log domains around and protect them by an additional mutex
to not corrupt the log domains of handlers running in other threads.
When setting the handler glib handles that for us since handler
replacement is already mutex protected.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>