osk-manager: Handle visibility property directly
This makes the visibility property r/o and makes sure
it gets updated when the corresponding DBus property actually changes
(after the async call or when we're told it changed from the OSK).

This makes sure they stay in sync. This is also simpler since we don't
need to bother with variant parsing, etc.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>