Add app-tracker
This keeps track of spawned apps (via glib's g_app_info) and
corresponding startup_id's (mostly via gtk_shell1 protcol).

Additionall we track org.gtk.gio.DesktopAppInfo on DBus

This allows us to show splash screens for apps that are spawned by other
processes than the shell itself. E.g.

    gtk4-launch gnome-control-center.desktop

or when opening a file in nautilus. We only do something if we get a
desktop file name and a startup id (the later is e.g. not the case when
processes are spawned via 'gio launch' or gnome-contacts).

Signed-off-by: Guido Günther <guido.gunther@puri.sm>