screen-saver-manager: Separate locking from blanking
Honor org.gnome.desktop.screen-saver lock-enabled so we don't lock the
screen if that is disabled and handle lock-delay to not lock
immediately.

For that we mark screensaver as active/inactive based on the power mode
of the primary monitor. We don't use the built-in monitor since that can
be disabled in multi-screen setups but the primary is the last one to go
off.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1012>