docked-manager: Check whether docked mode can be enabled
For some reason, docked mode sometimes gets enabled on its own (#421).
The reason why that happens is still unknown.

When that happens while the user has no dock around and no way to
easily ssh into the phone, the only thing that they can do in order
to make the phone fully functional again is to reboot it - otherwise
there's no OSK available and windows don't get maximized. Therefore,
let's log an error when that happens in hopes that it will provide
at least some insight about the issue.