gitlab-ci: Use gcovr instead of lcov
This creates a bit nicer reports but more importantly supports
a gcovr.cfg in _build/ so we can exclude autogenerated files
in the future.

This needs some hackery due to
https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=987818
but since this only affects this project and not e.g. downstreams
it should be o.k. for the time being.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>