rotation-manager: Cancel async operation on dispose
Instead of holding a ref during the async operation we cancel it in
dispose. Otherwise we might not properly dispose the object while an
async operation is pending.

We shuffle the code around to not access self prior to checking if the
operation was canceled since it might not be existent anymore.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>