end-session-dialog: Make listbox insensitive
The list of inhibited apps should just be avoided when using keyboard
focus however just setting can-focus = False on the listbox and the
rows only makes it swallow the cursor completely. Mark the widget
as insenitive instead. This dims the widgets in the listbox so
undo that via CSS.

With that using tab to cycle through the dialog works as expected (only
switches bestween the buttons).

Signed-off-by: Guido Günther <guido.gunther@puri.sm>