tests: Always run idle-manager test
Now that we can use PhoshShell in tests use that
in a thread plus a spawned compositor to always
run the idle tests instead of requiring a running
shell upfront.

For that we need to fix resource leaks to not reuse
DBus proxys on already gone shells, etc.

We mark DBus session bus using tests as not being run in parallel for
now since test clients might otherwise get confused.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>