shell: Sort manager objects
Sort the ones that get created by the shell and those created on the fly
into separate sections and sort those alphabetically. Ideally we want
to distingiush the getters by name at some point.

(There's a third type (singletons) which we dont' bother about here and
which we're reducing bit by bit)

Signed-off-by: Guido Günther <guido.gunther@puri.sm>