gitlab-ci: Run extended tests on bullseye
Don't bother for the crossbuild since we don't run tests there and this
would just make build times longer.

Since we dont't want to build documentation for glib we need to do an
extra build for gtk-doc that uses system glib.

This gets simpler again once the next glib is released.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>