Move feedback for notifications to a separat class
It was o.k. to have it in settings since that was the only long term
notification store and we just used a single feedback event with

- notifications also being on the lock screen
- sending different event feedback

we don't want to add more complexity to that single notification
display. This also stops abusing the ListBox's create_row callback to
handle feedback events.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>