notify-feedback: Adjust event name based on notification category
Adjust the emitted feedback based on the notifaction category¹. We could
use the sound-name instead but that has the problem that the application
sometimes has a hard time to figure out what the appropriate event is.
E.g. `message-missed-instant` and `message-new-instant` are hard to tell
apart from an application PoV.

1: https://specifications.freedesktop.org/notification-spec/notification-spec-latest.html

Signed-off-by: Guido Günther <guido.gunther@puri.sm>