app-tracker: Adjust to "launch-started" DBus API
GLib 2.72 emits "launch-started" for spawned and DBus activated apps.
Track this too and use that instead of "launched" to bring up the splash
since the later is emitted only once the app is up which is too late for
a splash.

There's some backward compatibility code to try get reasonable results
with current released glib but for DBus acivated apps you really need
2.72.0.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>