mode-manager: Handle virtual monitor
If a system has a single output and that is virtual we treat it as not
having external display hardware. This allows with virtual wayland
backends like headless, x11 or wayland to test the undocked case more
easily.

If we want to test docked mode we can add a second output.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>