shell: Update builtin monitor in on_monitor_removed as well
So far we only unset the monitor but that's not enough if there is
another builtin monitor that we can use. This can be the case if e.g. a
phone's internal DSI-1 gets disabled and later on reenabled.

We do this by moving setting and unsetting of the builtin monitor to
phosh_shell_set_builtin_monitor() which makes it more symmetric with
phosh_shell_set_primary_monitor() and also moves the signal emission to
a single function so we only get a single `notify::builtin-monitor` when
it gets updated..

Signed-off-by: Guido Günther <guido.gunther@puri.sm>