tests: Fix translations for call related screen shots
libcall-ui has no idea we're picking the translations from odd
places so help it along by setting the path to the translations
again after cui_init().

We can't do that in main since the creation of the shell object
happens after in the fixture (and we don't want to do that in
the fixture since we don't want to leak test details of the
screenshot tests into testlib).

Signed-off-by: Guido Günther <guido.gunther@puri.sm>