app-grid-button: Limit the natural width request
This uses a PhoshClamp to limit the natural width request of the button
in pixels rather than limiting the label's maximum width in characters.
This dissociates the width from the font size, giving us more control
over the button's size, and this allows the button's labels to take more
width if available.

This purposefully leaves the indentation broken to ease review. It will
be fixed in the next commit.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1002>