Emit WakeUpScreen signal and reset idle timers on resume
When the device resumes from suspend and there is no user input,
gsd-power remains in the "sleep" state because there is no WakeUpScreen
signal such that even when the new idle timer that gsd-power registered
after resume triggers, the transition from "sleep" to "sleep" is
ignored and the system does not suspend. A second difference to the
behavior of Mutter/GNOME Shell is that on resume there is no reset of
the registered idle timers which should also be fixed to ensure that
Phosh behaves the same.

Act on the logind PrepareForSleep signal (the resume case) to emit the
WakeUpScreen signal and reset the idle timers.

Fixes https://gitlab.gnome.org/World/Phosh/phosh/-/issues/686

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1016>