docs: Drop wayand protocols
Since they're not using gtk-doc parsing always gave heaps of errors and
the generated don't have much to say either.

It would be great to have them documented (e.g. by linking to the
upstream docs) but we can do that with the switch to gi-docgen.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1027>