lockscreen: Move call-ui down by top-panel's height
Since the top-bar now takes up space we don't want the back
button to overlap the top-bar.

We chose that approach instead of changing the lockscreens exclusive
zone so that it keeps to extend below the top bar. This will make e.g. a
transparent top bar work as the background can shine through and also
make the "centered" property of GtkBoxes on the lock screen still center
on the screen without having to care for additional padding.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1048>