shell: Hide top-panel when proximity fader is on
Since the top-panel uses an exclusive zone and the fader doesn't, the
top-panel is arranged on top of the fader. Due to this the top-panel
would still accept input e.g. during phone calls which leads to odd
results when typing with the ear.

So hide the top-panel for the moment when the fader is active. Later on
we can move it one layer down which is a bit nicer but needs the
corresponding phoc patches to support that.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1048>