shell: Lower top-panel when proximity fader is on
Since the top-panel uses an exclusive zone and the fader doesn't the
top-panel is arranged on top of the fader. Due to this the top-panel
would still accept input e.g. during phone calls which leads to odd
results when typing with the ear.

Moving the top-panel to another layer is a bit nicer than just hiding
it since we still have the layer around on the unlocked shell in
case something goes wrong.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1053>