top-panel: Use the full screen height for the top panel
For that we reduce the notification's scrolled window minimum height a
bit as it will expand by itself.

This prevents the settings menu from expanding beyond the lower screen
edge in portrait mode (which happend so far when media player and
notifications were shown).

Using the fulls screen height is also in line with the designs -
although it will be a bit black heavy in docked mode. We will need to
compensate for that e.g. by using less black (and e.g. just blur the
underlying screen space).

This is in preparation for the gesture code as the dragable surface also
uses the full screen hight.

TODO: honor mode change

Signed-off-by: Guido Günther <guido.gunther@puri.sm>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1057>