shell: Honor modal dialogs and full screen when raising top-bar
Just handling the fader isn't enough as we want full screen windows and
modal dialogs to stretch over the top bar as well.

We have some more cases like screenshost screen flash but that needs
proper z-level which we can introduce once the layer-shell-effects is
in. Same if we want system modal dialogs on the lock screen.

Fixes: 55bbe5758ba947a1d77c4b75d4424ea8cf4ebac6
Closes: https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/370

Signed-off-by: Guido Günther <guido.gunther@puri.sm>