shell: Set monitor when creating rotation manager
I had an annoying bug on Samsung phones such as GT-I9070,
GT-I8160 and GT-I8530 all using DPI displays that did not
properly instantiate the rotation manager, while
phones on the same platform using DSI would work fine.

If the primary monitor has already arrived when the rotation
manager is instantiated, the monitor does not get assigned to
the rotation manager.

This happens of phones such as Samsung GT-I9070 which use
DPI displays on MCDE that arrive apparently very quickly.
The same U8500 platforms with DSI displays does not have this
problem, presumably because the primary monitor arrives
later.

This fix was developed working entirely on the target,
compiling Phoc and Phosh natively on the GT-I9070 dual Cortex-A9
and iteratively editing, recompiling and testing modifications
to Phosh on the target itself.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>