overview: Don't update thumbnails when overview is unfolded
We don't do live updates of thumbnails yet, so having them update out of
sudden but only in some situations makes it look unpolished. We still want
to update when overview is folded in order to keep cache of thumbnails
in state that was seen by the user the last time when windows change focus.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1085>