overview: Raise toplevels on activity scrolling
Since the user can now swipe the home screen down to dismiss it,
doing so instead of tapping on an activity feels more like an
alternate way of getting the home screen out of the way instead
of canceling what you were doing there. This makes an interaction
where the user scrolls through activity list and then swipes it
away feel weird, as the user would be moved back to the previously
active application instead of the one they just scrolled to.

Fix that by activating the toplevel right away when the user
scrolls to it. This doesn't actually bring focus to the window,
as it's stolen by phosh's layer-shell surface due to it being
keyboard interactive, but it does raise it up and makes it
active once the user folds the home screen.

Do it only when not docked, as raising windows when you scroll
through the list is not really the expected behavior on multi
screen setups.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1097>