lockscreen: Focus PIN entry on carousel page change when OSK is on
In a case where the OSK would stay visible behind the lockscreen,
we would end up with no numpad *and* no focus on the PIN entry,
so the OSK wouldn't get elevated above the lockscreen and the user
would have no way to input their password other than using a
physical keyboard.

The easiest way to reproduce this scenario is to run:

`killall squeekboard; SQUEEKBOARD_DEBUG=force-show squeekboard`

while the screen is locked.

Fix that by ensuring that the PIN entry gets focused as soon
as the carousel is switched to the unlock page whenever
the keypad is hidden.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1110>