shell-manager: Report version
GNOME Settings fetches the version via DBus so report a version number
but make it very obvious it's not the pure GNOME version.

We can use GNOME_PLATFORM_VERSION to report a meaningful GNOME release
we base on for GNOME > 43. If that's not available fall back to parsing
the gnome-desktop XML (as gnome-settings did for GNOME < 43). We can
drop that bit once we can assume GNOME 43 as baseline.

Thanks to Jeremy Bicha for bringing this up.

Signed-off-by: Guido Günther <guido.gunther@puri.sm>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1120>