Update to gmobile 0.0.1
This is 0.0.1 plus a few patches to build with older glib and to build
json-glib as a fallback when not present.

We adjust to the new API in the same commit as it's a minor change
and keeps things bisectable.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1196>