notify-feedback: Handle locked as blanked
When the screen is locked and something lights it up (or the user is
using a plugin) we want to make sure there's some feedback that a new
event arrived so handle lock as blanked for now.

We'll revisit this when continuing the notification feedback cleanup.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1207>