screen-saver-manager: Only blank on key release
This avoids unblanking the screen if the user presses the button for a
longer period of time. It will also allow us to handle long press.

Unblanking already happens on power button press as the activity fed by
the compositor is enough (similar to e.g. the volume buttons).

Closes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/911

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1220>