emergency-calls: Disable by default
We want to make sure the code is working reliably. For that we
need a stable gnome-calls release and testing among more providers.
Hence toggle it off by default atm until that happened. It can
be enabled anytime using:

  gsettings set sm.puri.phosh.emergency-calls enabled true

Distributions that want to enable it by default can use an override.

We do this in separate commit so that we can easily revert it
once we want to enable by default.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/904>