home: Handle the super key
Since Phoc allows passing the super key (see MR) we can subscribe to it
and use the super key to unfold the home. This is similar to how the key
works on Gnome.

Note that on versions of Phoc that don't expose the super key, this will
just generate a warning as Phoc rejects the binding.

Works-with: https://gitlab.gnome.org/World/Phosh/phoc/-/merge_requests/447
Fixes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/563
Signed-off-by: Alistair Francis <alistair@alistair23.me>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1236>