rotation-manager: Match up orientation after unlock
We claim the accelerometer on unblank and make the orientation match the
sensor readings. However when the device is locked we ignore these
values. For these cases we want to match the orientation again after
unlock.

Thanks: Adam Plumb for the suggestion

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1240>