settings: Allow to select audio devices
This can be useful to switch between headsets and speaker without having
to open g-c-c.

For that we introduce PhoshAudioDevices which is a list model to either
store input or output devices. These are then put into a GtkListBox that
is presented by PhoshAUdioSettings. PhoshAuidoSettings keeps the audio
input and output related information in PhoshSettings together.

Closes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/951

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1277>