call-notification: Make sure we always show a phone number when available
For calls that don't have an entry in the address book we want to make
sure we *only* show the phone number and nothing else (not e.g. "Unknown
Caller") while the reset of the logic (whether to show initials, etc)
should stay the same.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1288>