build: Simplify libs
We build a static lib for use with the phosh binaries, tests, etc and a
dynamic one for the tests but don't need an extra dependency.

Both libs use a common set of headers and if we link the main library we
don't need to link with tools too.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1361>