status-page: Create PhoshStatusPage class
A status page is used in the quick settings to display more information
about a setting. For example, Wi-Fi quick setting can use it to show
available networks.

It must be subclassed accordingly and the header widget must be set as needed.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1303>