settings: Show Wi-Fi status page on long pressing Wi-Fi quick setting
By this commit we add WifiStatusPage as a child to a stack. The flowbox
of PhoshQuickSettings is now a child of this stack. When a quick setting
is long pressed, the status page can be shown.

We also add a new action to close the status page.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1303>