end-session-dialog: Use cancellable
As in other places properly cancel async calls during shutdown so we
don't need to hold a ref on self (which can then again trigger invalid
access as widgets are already disposes).

This also avoids a leak of self in the error path.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1322>