ci: Add Uncrustify to Debian image
`uncrusify` is a utility to check the code style of C, Vala etc. files.
We update the `debian.Dockerfile` to include `uncrusify` and add the
updated the image to be used in GitLab CI.
Similarly, `screenshot.Dockerfile` is updated to include the updated
Debian image.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1327>