lockscreen: Add small margin to the widget box
If there are long filenames in the ticket box it expands the widget box
right to the screen edge breaking the swipe back.

By adding a small margin we always allow a fling at the screen edge to
go back to the actual lockscreen.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1337>