phosh-session: Reduce number of processes
The phoc -E switch uses "sh -c" to run the given commands and then in
that shell "bash -lc" is started to run the session. Without using
"exec CMD" this leaves the shell process around while they wait for the
child process.
To save memory, replace the shell process in the second invocation, but
leave the first as is to address this in phoc.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1352>