background: Load images async
Instead of handling loading in the background (and potentially blocking
drawing) load them async. For that we move the settings handling into
the background manager (which also handles clearing the cache on
settings changes).

As the background now calls back into the background-manager we finish
background-manager's setup in an idle callback to ensure the background
-manager is fully up when the background needs it.

This simplifies the background handling as the background-manager no
longer needs to be concerned with the details of the layer surface.

Closes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/447
Closes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/910
Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1356>