home: Clarify home-bar name
It's not only a button since some time so let's use consistent naming.
This also clarifies the difference between the home-bar (the bar at the
bottom of the screen) and the home widget which contains the bar and the
overview.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1340>