top-panel: Use a background
This allows us to tweak the top-panel's transparency based on how far
the surface is folded.

We use that to make it fully transparent (and hence the wallpaper shine
through) when the overview is unfolded or when on the lockscreen.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1262>