test-take-screenshots: Take the display out of power save mode
When locking the screen the output is put into power saving mode. To
wake it up again the ScreenSaverManager sends "WakeupScreen" which is
caught by g-s-d.  However we can't assume g-s-d to be available in the
tests so make sure the screen is powered as we can otherwise not take
any screenshots.

With out this we might fail with:

   (.../phosh/_build-test/tests/test-take-screenshots:660974): phosh-screenshot-manager-WARNING **: 14:10:26.777: Failed to copy output 'HEADLESS-1'

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1364>