screen-saver-manager: Use lock-enabled in more places
Except in cases where the lock is requested explicity (like the DBus
call) we want to take the lock-enabled setting into account.

This fixes the case where lock-enabled is false but a lock-delay is set.
We don't want to lock in that case either.

Fixes: 1a2c6dbb1 ("screen-saver-manager: Split arming the lock delay timer into separate function")

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1371>