build: Split dbus header and source targets
This allows us to only use the header dependencies for the plugins as
there's no point in generating the sources (which would then happen for
every plugin).

We only split the ones that already use our default namespace as we want
to fix up the namespaces for the other ones first as this gives shorter
names. See e.g.

   aa6c9323b ("session-manager: Use default DBus prefix")

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1373>