plugins: Screen scaling quick setting plugin
Allow to select the scale for the primary output. The list of scales
is calculated the same way we do it when presenting them to Settings.

For the moment we disable the plugin when there's more than one output
as it's not obvious to the user if it should affect the primary or
built-in display.

Clicking on the button swaps back to the previously selected scale which
looks like the best default action.

Co-authored-by: Guido Günther <agx@sigxcpu.org>
Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1400>