panels: Queue draw instead of commiting the surface
If we commit too early and GTK hasn't updated the buffer the compositor
will terminate us due to the buffer size not matching the scale.

A draw will eventually commit the surface. This makes sure we don't
commit without giving GTK a chance to update e.g. the buffer size.

This is not ideal as we invalidate the complete surface causing a full
redraw.

Closes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/1050
Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1405>