app-grid: Find the valid button to focus by loop
Previously we focused either the last button or the second last button.
But this resulted in issues when multiple apps are removed. So to be on
safer side, find the valid button to focus by looping in
reverse. (Assuming that when main grid is shown again, it has at least
one button.)

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1407>