app-grid: Add ability to edit folder names
We add an entry that is shown when the edit button is toggled to active. When user
toggles the button again, the entry's contents are set as folder name.
Pressing Enter (or any activation) with the entry focused is same as
toggling button again.
Closing the folder (by pressing Escape or back-button or gesture) or
setting the entry to empty string doesn't do any changes, that is,
cancels the edit.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1407>