folder-info: Reload apps when settings changes
Previously, whenever GSettings' `app` key changes, we emit
`apps-changed` signal, so that app-grid and other dependents can regroup
apps if needed.

This had one issue. Since we don't reflect the changes inside
folder-info, if the folder is open, then the user doesn't see the new
apps. They had to close and reopen the folder, to see new apps.

In this commit, we reload the apps whenever the key changes in-addition
to signal emission.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1407>