top-panel: Set initial shell state to unfolded
This is an artifact of the fact that the drag surface code needs some
time to wiggle the surface into place (which we want to fix on the
compositor side). Once that is fixed we can then revert this patch.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1413>