lockscreen: Allow custom unlock submit callback
The callback that is fired when a passphrase is submitted from the
keypad entry is virtualized so that subclasses can override this
behaviour.

This change primarily benefits Phosh-based greeters like Phrog. In this
case, the login behaviour is entirely different from the default (the
PAM conversation is mediated through greetd IPC, and takes place before
there's actually any user logged in).

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1437>