lockscreen: Always clear idle timer in carousel_position_notified_cb
Even if the keypad is already visible, any swiping activity triggers
carousel_page_changed_cb again anyway.

This actually slightly improves UX in a somewhat edge case condition. If
you swipe to the keypad, and then swipe up/down in the keypad region
without any keypad input (or switching all the way back to the page
above), the idle timer will kick in and suddenly take control of the
carousel. With this change, *any* swiping behaviour properly resets the
idle timer.

This is fortunate, because the intent is to remove "PhoshLocksreenPos"
imminently, and this change removes two usages of that enum.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1437>