lockscreen: Allow an "extra" page to be inserted
If phosh_lockscreen_add_extra_page is called, the carousel will have an
extra page inserted between the info page and the unlock page. This
extra page can then be navigated via swipes, and also set as the
current and/or default page.

This is useful for the Phosh-based greeter use case (for example Phrog),
which inserts a user/session selection page.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1437>