power-menu-manager: Allow to suspend from power menu
If suspend is allowed we show the suspend action rather then disabling
the power off button.

This makes the power button menu more useful on the lock screen. And
since it's a non-destructive action there's no issue with allowing it.

Closes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/990
Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1438>