status-page: Add placeholder / empty state widget
Add a placeholder widget for the displaying the empty/disabled state in
settings status pages. This is very similar to `HdyStatusPage` but we
add a custom widget as we can't scale down the icons in GTK3 via CSS as
-gtk-icon-size doesn't exist yet and -gtk-icon-scale scales the icon but
doesn't shring the size allocations.

This also helps to ease displaying empty states for list boxes as the
`placeholder` there doesn't work well with advanced styling. See
libhandy#468.

This can be replaced with AdwStatusPage once we switch to GTK4.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1469>