bt-manager: Replace #include <bluetooth-device.h> with a typedef
This header is vendored into phosh because upstream does not want to
make it public, so we can't leak it into libphosh headers.

Further usage of types from these vendored headers are not anticipated
for now, so we can get away with a typedef for the one BluetoothDevice
type we need.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1491>