wwan: Improve doc strings
So far we only documented PhoshWWanInterface but not PhoshWWan so it
wouldn't show up. While at that modernize some other doc strings around
wwan and add since annotations to the classes as it can be helpful to
know how old/mature an implementation is.

Gbp-Dch: Ignore
Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1504>