top-panel: Set background on top bar
This avoids showing regular windows under the top bar which can be
confusing.

While this is not very visible in auto-maximize mode it can be seen in
floating mode or when e.g. revealing the shell over fullscreen windows.

We add an outer container with background to the top bar so we can keep
setting the margins on the top-bar itself.

The alternatives (using padding or to setting the marign on the inner
elements) are more complicated as we move elements around due to notch
avoidance and the margin also changes due to that. We would thus need to
generate style classes on the fly or determine the elements dynamically
which is more fragile.

We can make things slightly nicer by setting the solid
background a bit later (when the splash closes) but let's do that with
when zooming in the splash
(https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1264)
as this will supply the necessary plumbing to phosh shell.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1597>