metainfo-cache: Introduce PhoshMetainfoCache
This was originally authored by Guido in the WIP MR !950.

It handles initializing an AppStream cache "Pool", which is populated
with information about available applications on the system (installed
or otherwise).

It will be utilized in the following commit of this series, to determine
the correct AppStream "data id" that is used when invoking
gnome-software --uninstallation

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1600>